+++ src/Common/Database/AbstractDatabase.php
@@ -20,7 +20,7 @@ use ApacheBorys\Location\Model\Place;
 /**
  * @author Borys Yermokhin <borys_ermokhin@yahoo.com>
  */
-abstract class AbstractDatabase
+abstract class AbstractDatabase implements DataBaseInterface
 {
     /**
      * @var bool[]
@@ -106,8 +106,6 @@ abstract class AbstractDatabase
      * Levels compiler for forming identifier for Address entity in @see compileKey
      *
      * @return string[]
-     *
-     * @throws \Psr\Cache\InvalidArgumentException
      */
     protected function compileLevelsForKey(Address $address): array
     {
@@ -190,22 +188,35 @@ abstract class AbstractDatabase
      * @param int    $maxResults
      * @param string $locale
      * @param int    $filterAdminLevel
+     * @param string $filterType
      *
      * @return string[]
      */
-    protected function makeSearch(string $phrase, int $page, int $maxResults, string $locale, int $filterAdminLevel = -1): array
-    {
+    protected function makeSearch(
+        string $phrase,
+        int $page,
+        int $maxResults,
+        string $locale,
+        int $filterAdminLevel = -1,
+        string $filterType = ''
+    ): array {
         $result = [];
 
-        foreach ($this->actualKeys[$locale] as $adminLevel => $actualKeys) {
-            if ($filterAdminLevel > -1 && $filterAdminLevel !== $adminLevel) {
+        foreach ($this->actualKeys[$locale] as $type => $adminLevels) {
+            if (strlen($type) > 0 && $type !== $filterType) {
                 continue;
             }
 
-            foreach ($actualKeys as $actualKey => $objectHash) {
-                $grade = $this->evaluateHitPhrase($phrase, $actualKey);
-                if ($grade > 0) {
-                    $result[$actualKey] = $grade;
+            foreach ($adminLevels as $adminLevel => $actualKeys) {
+                if ($filterAdminLevel > -1 && $filterAdminLevel !== $adminLevel) {
+                    continue;
+                }
+
+                foreach ($actualKeys as $actualKey => $objectHash) {
+                    $grade = $this->evaluateHitPhrase($phrase, $actualKey);
+                    if ($grade > 0) {
+                        $result[$actualKey] = $grade;
+                    }
                 }
             }
         }
@@ -256,13 +267,15 @@ abstract class AbstractDatabase
      * @param string $locale
      * @param string $key
      *
-     * @return int
+     * @return array
      */
-    protected function findAdminLevelForKey(string $locale, string $key): int
+    protected function findAdminLevelAndTypeForKey(string $locale, string $key): array
     {
         foreach ($this->existAdminLevels as $existAdminLevel => $value) {
-            if (isset($this->actualKeys[$locale][$existAdminLevel][$key])) {
-                return $existAdminLevel;
+            foreach (array_keys($this->actualKeys[$locale]) as $existType) {
+                if (isset($this->actualKeys[$locale][$existType][$existAdminLevel][$key])) {
+                    return [$existAdminLevel, $existType];
+                }
             }
         }