+++ src/Common/Database/Psr6Database.php
@@ -14,6 +14,7 @@ namespace ApacheBorys\Location\Database;
 
 use ApacheBorys\Location\Model\DBConfig;
 use ApacheBorys\Location\Model\Place;
+use ApacheBorys\Location\Model\PlaceCollection;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\InvalidArgumentException;
 
@@ -25,11 +26,12 @@ class Psr6Database extends AbstractDatabase implements DataBaseInterface
     /**
      * By that keys we will store hashes (references) to fetch real object
      * first key - locale
+     * second key - type of Place
      * second key - admin level
      * third key - compiled key from Address of current locale
      * value - hash of object
      *
-     * @var string[][][]
+     * @var string[][][][]
      */
     protected $actualKeys = [];
 
@@ -123,9 +125,9 @@ class Psr6Database extends AbstractDatabase implements DataBaseInterface
         $result = [];
 
         foreach ($this->makeSearch($searchKey, $page, $maxResults, $locale, $filterAdminLevel) as $key) {
-            $adminLevel = $this->findAdminLevelForKey($locale, $key);
+            list($adminLevel, $placeType) = $this->findAdminLevelAndTypeForKey($locale, $key);
 
-            $item = $this->databaseProvider->getItem($this->actualKeys[$locale][$adminLevel][$key]);
+            $item = $this->databaseProvider->getItem($this->actualKeys[$locale][$placeType][$adminLevel][$key]);
             if ($item->isHit()) {
                 $this->dbConfig->isUseCompression() ?
                     $rawData = json_decode(gzuncompress($item->get()), true) :
@@ -143,10 +145,10 @@ class Psr6Database extends AbstractDatabase implements DataBaseInterface
      *
      * @throws \Psr\Cache\InvalidArgumentException
      */
-    public function getAllPlaces(int $offset = 0, int $limit = 50): array
+    public function getAllPlaces(int $offset = 0, int $limit = 50): PlaceCollection
     {
         if ($offset > count($this->objectsHashes)) {
-            return [];
+            return new PlaceCollection();
         }
 
         if ($limit > $this->dbConfig->getMaxPlacesInOneResponse()) {
@@ -177,7 +179,7 @@ class Psr6Database extends AbstractDatabase implements DataBaseInterface
             }
         }
 
-        return $result;
+        return new PlaceCollection($result);
     }
 
     /**
@@ -198,8 +200,8 @@ class Psr6Database extends AbstractDatabase implements DataBaseInterface
         foreach ($this->actualKeys as $locale => $keys) {
             $place->selectLocale($locale);
             $keyForDelete = $this->compileKey($place->getSelectedAddress());
-            if (isset($keys[$place->getMaxAdminLevel()][$keyForDelete])) {
-                unset($this->actualKeys[$locale][$place->getMaxAdminLevel()][$keyForDelete]);
+            if (isset($keys[$place->getType()][$place->getMaxAdminLevel()][$keyForDelete])) {
+                unset($this->actualKeys[$locale][$place->getType()][$place->getMaxAdminLevel()][$keyForDelete]);
                 $this->updateActualKeys();
             }
         }
@@ -356,7 +358,7 @@ class Psr6Database extends AbstractDatabase implements DataBaseInterface
         $this->updateHashKeys();
 
         foreach ($this->compileKeys($place) as $locale => $key) {
-            $this->actualKeys[$locale][$place->getMaxAdminLevel()][$key] = $place->getObjectHash();
+            $this->actualKeys[$locale][$place->getType()][$place->getMaxAdminLevel()][$key] = $place->getObjectHash();
         }
         $this->updateActualKeys();